@@ -8,10 +8,10 @@ publish_to: "none"
 # followed by an optional build number separated by a +.
 # Both the version and the builder number may be overridden in flutter
 # build by specifying --build-name and --build-number, respectively.
-version: 0.8.6
+version: 0.9.2
 
 environment:
-  sdk: ">=2.18.2 <3.0.0"
+  sdk: ">=2.19.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -121,11 +121,11 @@ dependencies:
   pretty_qr_code: ^2.0.2
   dotted_border: ^2.0.0+2
   image_picker: ^0.8.6
-  mobile_scanner: ^3.0.0-beta.4
+  mobile_scanner: ^3.4.1
+  # mobile_scanner: ^3.0.0-beta.4
   # mobile_scanner: 3.0.0-beta.1
   # Share files
   share_plus: ^7.0.0
-
 
   # File picker
   file_picker: ^5.2.0
@@ -204,7 +204,7 @@ dependencies:
       url: https://github.com/perishllc/flutter_sodium
       ref: main
   x25519: ^0.1.1
-  encrypt: ^5.0.1
+  encrypt: ^5.0.3
 
   # search bar + text highlighting:
   # searchbar_animation: ^0.0.3
@@ -355,6 +355,8 @@ flutter:
     # illustrations:
     - assets/illustrations/transferfunds_start.svg
     - assets/illustrations/transferfunds_paperwalletonly.svg
+    # nano-pow:
+    - assets/work/
     # changelog:
     - assets/CHANGELOG.md
 
@@ -362,3 +364,4 @@ flutter:
   # included with your application, so that you can use the icons in
   # the material Icons class.
   uses-material-design: true
+  generate: true
