@@ -102,7 +102,7 @@ class MessageLookup extends MessageLookupByLibrary {
         "blockUser":
             MessageLookupByLibrary.simpleMessage("Bloker denne bruger"),
         "blockedAdded":
-            MessageLookupByLibrary.simpleMessage("% 1 blev blokeret."),
+            MessageLookupByLibrary.simpleMessage("%1 blev blokeret."),
         "blockedExists":
             MessageLookupByLibrary.simpleMessage("Bruger allerede blokeret!"),
         "blockedHeader": MessageLookupByLibrary.simpleMessage("blokeret"),
@@ -115,7 +115,7 @@ class MessageLookup extends MessageLookupByLibrary {
         "blockedNameMissing":
             MessageLookupByLibrary.simpleMessage("Vælg et kaldenavn"),
         "blockedRemoved":
-            MessageLookupByLibrary.simpleMessage("% 1 er blevet blokeret!"),
+            MessageLookupByLibrary.simpleMessage("%1 er blevet blokeret!"),
         "branchConnectErrorLongDesc": MessageLookupByLibrary.simpleMessage(
             "Vi kan tilsyneladende ikke nå Branch API, dette skyldes normalt en form for netværksproblem eller VPN, der blokerer forbindelsen.\n\n Du bør stadig være i stand til at bruge appen som normalt, men at sende og modtage gavekort fungerer muligvis ikke."),
         "branchConnectErrorShortDesc": MessageLookupByLibrary.simpleMessage(
@@ -239,13 +239,14 @@ class MessageLookup extends MessageLookupByLibrary {
             "Vælg, hvilken enhed der skal vises beløb i.\n1 nyano = 0.000001 NANO, eller \n1,000,000 Nyano = 1 NANO"),
         "currentlyRepresented":
             MessageLookupByLibrary.simpleMessage("V současné době zastupuje"),
+        "daily": MessageLookupByLibrary.simpleMessage("Daglige"),
         "dayAgo": MessageLookupByLibrary.simpleMessage("For en dag siden"),
         "decryptionError":
             MessageLookupByLibrary.simpleMessage("Dekryptering Fejl!"),
         "defaultAccountName":
             MessageLookupByLibrary.simpleMessage("Hlavní účet"),
         "defaultGiftMessage": MessageLookupByLibrary.simpleMessage(
-            "Tjek Nautilus ud! Jeg sendte dig noget nano med dette link:"),
+            "Tjek Nautilus ud! Jeg sendte dig noget %2 med dette link:"),
         "defaultNewAccountName":
             MessageLookupByLibrary.simpleMessage("Účet %1"),
         "delete": MessageLookupByLibrary.simpleMessage("Slet"),
@@ -376,7 +377,7 @@ class MessageLookup extends MessageLookupByLibrary {
         "favoriteNameMissing": MessageLookupByLibrary.simpleMessage(
             "Vælg et navn til denne favorit"),
         "favoriteRemoved": MessageLookupByLibrary.simpleMessage(
-            "% 1 er blevet fjernet fra favoritter!"),
+            "%1 er blevet fjernet fra favoritter!"),
         "favoritesHeader": MessageLookupByLibrary.simpleMessage("favoritter"),
         "featured": MessageLookupByLibrary.simpleMessage("Fremhævede"),
         "fewDaysAgo": MessageLookupByLibrary.simpleMessage("For få dage siden"),
@@ -395,6 +396,8 @@ class MessageLookup extends MessageLookupByLibrary {
         "fundingBannerHeader":
             MessageLookupByLibrary.simpleMessage("Finansieringsbanner"),
         "fundingHeader": MessageLookupByLibrary.simpleMessage("Finansiering"),
+        "generatingWork": MessageLookupByLibrary.simpleMessage(
+            "Genererer PoW, dette kan tage et stykke tid (> 20 sekunder)..."),
         "getCurrency": MessageLookupByLibrary.simpleMessage("Hent %2"),
         "getNano": MessageLookupByLibrary.simpleMessage("Få NANO"),
         "giftAlert": MessageLookupByLibrary.simpleMessage("Du har en gave!"),
@@ -525,16 +528,17 @@ class MessageLookup extends MessageLookupByLibrary {
             MessageLookupByLibrary.simpleMessage("Meddelelse kopieret"),
         "messageHeader": MessageLookupByLibrary.simpleMessage("Zpráva"),
         "minimumSend": MessageLookupByLibrary.simpleMessage(
-            "Minimální částka pro odeslání je% 1 NANO"),
+            "Minimální částka pro odeslání je %1 %2"),
         "minuteAgo": MessageLookupByLibrary.simpleMessage("For et minut siden"),
         "mnemonicInvalidWord":
-            MessageLookupByLibrary.simpleMessage("% 1 není platné slovo"),
+            MessageLookupByLibrary.simpleMessage("%1 není platné slovo"),
         "mnemonicPhrase":
             MessageLookupByLibrary.simpleMessage("Mnemonická fráze"),
         "mnemonicSizeError": MessageLookupByLibrary.simpleMessage(
             "Tajná fráze může obsahovat pouze 24 slov"),
         "mnemonicSizeError2": MessageLookupByLibrary.simpleMessage(
             "Hemmelig sætning må kun indeholde 12 eller 24 ord"),
+        "monthly": MessageLookupByLibrary.simpleMessage("Månedlige"),
         "monthlyServerCosts": MessageLookupByLibrary.simpleMessage(
             "Månedlige serveromkostninger"),
         "moonpay": MessageLookupByLibrary.simpleMessage("MoonPay"),
@@ -578,7 +582,7 @@ class MessageLookup extends MessageLookupByLibrary {
         "notificationInfo": MessageLookupByLibrary.simpleMessage(
             "For at denne funktion skal fungere korrekt, skal meddelelser være aktiveret"),
         "notificationTitle":
-            MessageLookupByLibrary.simpleMessage("Přijato %1 NANO"),
+            MessageLookupByLibrary.simpleMessage("Přijato %1 %2"),
         "notificationWarning":
             MessageLookupByLibrary.simpleMessage("Notifikationer deaktiveret"),
         "notificationWarningBodyLong": MessageLookupByLibrary.simpleMessage(
@@ -724,7 +728,7 @@ class MessageLookup extends MessageLookupByLibrary {
         "removeBlocked":
             MessageLookupByLibrary.simpleMessage("Fjern blokeringen"),
         "removeBlockedConfirmation": MessageLookupByLibrary.simpleMessage(
-            "Er du sikker på at du vil fjerne blokeringen% 1?"),
+            "Er du sikker på at du vil fjerne blokeringen %1?"),
         "removeContact":
             MessageLookupByLibrary.simpleMessage("Odstranit kontakt"),
         "removeContactConfirmation": MessageLookupByLibrary.simpleMessage(
@@ -740,7 +744,7 @@ class MessageLookup extends MessageLookupByLibrary {
         "representatives": MessageLookupByLibrary.simpleMessage("Zástupci"),
         "request": MessageLookupByLibrary.simpleMessage("Anmodning"),
         "requestAmountConfirm":
-            MessageLookupByLibrary.simpleMessage("Anmodning% 1% 2"),
+            MessageLookupByLibrary.simpleMessage("Anmodning %1 %2"),
         "requestError": MessageLookupByLibrary.simpleMessage(
             "Anmodning mislykkedes: Denne bruger ser ikke ud til at have Nautilus installeret, eller har meddelelser deaktiveret."),
         "requestFrom": MessageLookupByLibrary.simpleMessage("Anmodning fra"),
@@ -774,6 +778,7 @@ class MessageLookup extends MessageLookupByLibrary {
         "retry": MessageLookupByLibrary.simpleMessage("Prøv igen"),
         "rootWarning": MessageLookupByLibrary.simpleMessage(
             "Zdá se, že vaše zařízení je \"rootováno\", \"jailbroken\" nebo upraveno způsobem, který ohrožuje zabezpečení. Před pokračováním se doporučuje resetovat zařízení do původního stavu."),
+        "save": MessageLookupByLibrary.simpleMessage("Gemme"),
         "scanInstructions": MessageLookupByLibrary.simpleMessage(
             "Naskenujte QR kód Nano \n adresy"),
         "scanNFC": MessageLookupByLibrary.simpleMessage("Send via NFC"),
@@ -810,7 +815,7 @@ class MessageLookup extends MessageLookupByLibrary {
             "Kan ikke anmode om fra mig selv"),
         "send": MessageLookupByLibrary.simpleMessage("Poslat"),
         "sendAmountConfirm":
-            MessageLookupByLibrary.simpleMessage("Odeslat %1 NANO"),
+            MessageLookupByLibrary.simpleMessage("Odeslat %1 %2"),
         "sendAmounts": MessageLookupByLibrary.simpleMessage("Send beløb"),
         "sendError": MessageLookupByLibrary.simpleMessage(
             "Došlo k chybě. Zkuste to později."),
@@ -949,9 +954,9 @@ class MessageLookup extends MessageLookupByLibrary {
         "transferClose": MessageLookupByLibrary.simpleMessage(
             "Klepnutím kamkoli zavřete okno."),
         "transferComplete": MessageLookupByLibrary.simpleMessage(
-            "%1 NANO úspěšně převedeno do vaší peněženky Nautilus. \n"),
+            "%1 %2 úspěšně převedeno do vaší peněženky Nautilus. \n"),
         "transferConfirmInfo": MessageLookupByLibrary.simpleMessage(
-            "Byla nalezena peněženka se zůstatkem %1 NANO. \n"),
+            "Byla nalezena peněženka se zůstatkem %1 %2. \n"),
         "transferConfirmInfoSecond": MessageLookupByLibrary.simpleMessage(
             "Prostředky převedete klepnutím na potvrzení. \n"),
         "transferConfirmInfoThird": MessageLookupByLibrary.simpleMessage(
@@ -961,7 +966,7 @@ class MessageLookup extends MessageLookupByLibrary {
         "transferHeader":
             MessageLookupByLibrary.simpleMessage("Převést finanční prostředky"),
         "transferIntro": MessageLookupByLibrary.simpleMessage(
-            "Tento proces přenese prostředky z papírové peněženky do vaší peněženky Nautilus. \n\n Začněte klepnutím na tlačítko \"% 1\"."),
+            "Tento proces přenese prostředky z papírové peněženky do vaší peněženky Nautilus. \n\n Začněte klepnutím na tlačítko \"%1\"."),
         "transferIntroShort": MessageLookupByLibrary.simpleMessage(
             "Denne proces overfører midlerne fra en papirpung til din Nautilus-tegnebog."),
         "transferLoading": MessageLookupByLibrary.simpleMessage("Přenáší se"),
@@ -1034,6 +1039,7 @@ class MessageLookup extends MessageLookupByLibrary {
         "watchOnlySendDisabled": MessageLookupByLibrary.simpleMessage(
             "Afsendelser er deaktiveret på vagtadresser"),
         "weekAgo": MessageLookupByLibrary.simpleMessage("En uge siden"),
+        "weekly": MessageLookupByLibrary.simpleMessage("Ugentlig"),
         "welcomeText": MessageLookupByLibrary.simpleMessage(
             "Vítejte v Nautilus. Nejprve můžete vytvořit novou peněženku nebo importovat stávající."),
         "welcomeTextLogin": MessageLookupByLibrary.simpleMessage(
