@@ -61,7 +61,7 @@ class MessageLookup extends MessageLookupByLibrary {
         "amountGiftGreaterError": MessageLookupByLibrary.simpleMessage(
             "הסכום המפוצל לא יכול להיות גדול מיתרת המתנה"),
         "amountMissing": MessageLookupByLibrary.simpleMessage("הזן סכום"),
-        "appWallet": MessageLookupByLibrary.simpleMessage("% 1 ארנק"),
+        "appWallet": MessageLookupByLibrary.simpleMessage("%1 ארנק"),
         "askSkipSetup": MessageLookupByLibrary.simpleMessage(
             "שמנו לב שלחצת על קישור שמכיל קצת ננו, האם תרצה לדלג על תהליך ההגדרה? תמיד תוכל לשנות דברים מאוחר יותר.\n\n אם יש לך זרע קיים שברצונך לייבא עם זאת, עליך לבחור לא."),
         "askTracking": MessageLookupByLibrary.simpleMessage(
@@ -93,8 +93,7 @@ class MessageLookup extends MessageLookupByLibrary {
         "blockExplorerInfo": MessageLookupByLibrary.simpleMessage(
             "איזה Explorer לחסום להשתמש כדי להציג מידע העסקה"),
         "blockUser": MessageLookupByLibrary.simpleMessage("חסום משתמש זה"),
-        "blockedAdded":
-            MessageLookupByLibrary.simpleMessage("% 1 נחסם בהצלחה."),
+        "blockedAdded": MessageLookupByLibrary.simpleMessage("%1 נחסם בהצלחה."),
         "blockedExists":
             MessageLookupByLibrary.simpleMessage("המשתמש כבר חסום!"),
         "blockedHeader": MessageLookupByLibrary.simpleMessage("חסום"),
@@ -105,7 +104,7 @@ class MessageLookup extends MessageLookupByLibrary {
             MessageLookupByLibrary.simpleMessage("שם ניק כבר בשימוש!"),
         "blockedNameMissing":
             MessageLookupByLibrary.simpleMessage("בחר שם ניק"),
-        "blockedRemoved": MessageLookupByLibrary.simpleMessage("% 1 לא נחסם!"),
+        "blockedRemoved": MessageLookupByLibrary.simpleMessage("%1 לא נחסם!"),
         "branchConnectErrorLongDesc": MessageLookupByLibrary.simpleMessage(
             "נראה שאיננו יכולים להגיע ל-Banch API, זה בדרך כלל נגרם על ידי בעיה כלשהי ברשת או VPN שחוסם את החיבור.\n\n אתה עדיין אמור להיות מסוגל להשתמש באפליקציה כרגיל, אולם ייתכן ששליחה וקבלה של כרטיסי מתנה לא יפעלו."),
         "branchConnectErrorShortDesc": MessageLookupByLibrary.simpleMessage(
@@ -225,6 +224,7 @@ class MessageLookup extends MessageLookupByLibrary {
             "בחר באיזו יחידה להציג כמויות.\n1 ניאנו = 0.000001 ננו, או \n1,000,000 ניאנו = 1 ננו"),
         "currentlyRepresented":
             MessageLookupByLibrary.simpleMessage("מיוצג על ידי"),
+        "daily": MessageLookupByLibrary.simpleMessage("יום יומי"),
         "dayAgo": MessageLookupByLibrary.simpleMessage("לפני יום"),
         "decryptionError": MessageLookupByLibrary.simpleMessage("שגיאת פענוח!"),
         "defaultAccountName":
@@ -351,7 +351,7 @@ class MessageLookup extends MessageLookupByLibrary {
         "favoriteNameMissing":
             MessageLookupByLibrary.simpleMessage("בחר שם עבור מועדף זה"),
         "favoriteRemoved":
-            MessageLookupByLibrary.simpleMessage("% 1 הוסר מהמועדפים!"),
+            MessageLookupByLibrary.simpleMessage("%1 הוסר מהמועדפים!"),
         "favoritesHeader": MessageLookupByLibrary.simpleMessage("מועדפים"),
         "featured": MessageLookupByLibrary.simpleMessage("מובלט"),
         "fewDaysAgo": MessageLookupByLibrary.simpleMessage("לפני מספר ימים"),
@@ -367,6 +367,8 @@ class MessageLookup extends MessageLookupByLibrary {
         "fundingBannerHeader":
             MessageLookupByLibrary.simpleMessage("באנר מימון"),
         "fundingHeader": MessageLookupByLibrary.simpleMessage("מימון"),
+        "generatingWork": MessageLookupByLibrary.simpleMessage(
+            "יוצר PoW, זה עשוי לקחת זמן מה (> 20 שניות)..."),
         "getCurrency": MessageLookupByLibrary.simpleMessage("קבל %2"),
         "getNano": MessageLookupByLibrary.simpleMessage("קבל NANO"),
         "giftAlert": MessageLookupByLibrary.simpleMessage("יש לך כישרון!"),
@@ -488,8 +490,8 @@ class MessageLookup extends MessageLookupByLibrary {
             "תזכיר נשלח מחדש! אם עדיין לא הודה, ייתכן שהמכשיר של הנמען במצב לא מקוון."),
         "messageCopied": MessageLookupByLibrary.simpleMessage("ההודעה הועתקה"),
         "messageHeader": MessageLookupByLibrary.simpleMessage("הודעה"),
-        "minimumSend": MessageLookupByLibrary.simpleMessage(
-            "כמות השליחה המינימלית היא% 1"),
+        "minimumSend":
+            MessageLookupByLibrary.simpleMessage("כמות השליחה המינימלית היא%1"),
         "minuteAgo": MessageLookupByLibrary.simpleMessage("לפני דקה"),
         "mnemonicInvalidWord":
             MessageLookupByLibrary.simpleMessage("%1 is not a valid word"),
@@ -498,6 +500,7 @@ class MessageLookup extends MessageLookupByLibrary {
             "Secret phrase may only contain 24 words"),
         "mnemonicSizeError2": MessageLookupByLibrary.simpleMessage(
             "ביטוי סודי יכול להכיל רק 12 או 24 מילים"),
+        "monthly": MessageLookupByLibrary.simpleMessage("יַרחוֹן"),
         "monthlyServerCosts":
             MessageLookupByLibrary.simpleMessage("עלויות שרת חודשיות"),
         "moonpay": MessageLookupByLibrary.simpleMessage("MoonPay"),
@@ -674,7 +677,7 @@ class MessageLookup extends MessageLookupByLibrary {
             "אתה בטוח שאתה רוצה להחביא את החשבון? אתה יכול להוסיף אותו מחודש מאוחר יותר על ידי לחיצה על כפתור ה%1."),
         "removeBlocked": MessageLookupByLibrary.simpleMessage("בטל חסימה"),
         "removeBlockedConfirmation": MessageLookupByLibrary.simpleMessage(
-            "האם אתה בטוח שברצונך לבטל את החסימן% 1?"),
+            "האם אתה בטוח שברצונך לבטל את החסימן%1?"),
         "removeContact": MessageLookupByLibrary.simpleMessage("הסר איש קשר"),
         "removeContactConfirmation": MessageLookupByLibrary.simpleMessage(
             "האם אתה בטוח שאתה רוצה למחוק את %?"),
@@ -687,7 +690,8 @@ class MessageLookup extends MessageLookupByLibrary {
         "reply": MessageLookupByLibrary.simpleMessage("השב"),
         "representatives": MessageLookupByLibrary.simpleMessage("נציגים"),
         "request": MessageLookupByLibrary.simpleMessage("בקשה"),
-        "requestAmountConfirm": MessageLookupByLibrary.simpleMessage("בקשה% 1"),
+        "requestAmountConfirm":
+            MessageLookupByLibrary.simpleMessage("בקשה%1 %2"),
         "requestError": MessageLookupByLibrary.simpleMessage(
             "הבקשה נכשלה: נראה שלמשתמש זה לא מותקן נאוטילוס או שההודעות מושבתות."),
         "requestFrom": MessageLookupByLibrary.simpleMessage("בקשה מ"),
@@ -718,6 +722,7 @@ class MessageLookup extends MessageLookupByLibrary {
         "retry": MessageLookupByLibrary.simpleMessage("נסה שוב"),
         "rootWarning": MessageLookupByLibrary.simpleMessage(
             "It appears your device is \"rooted\", \"jailbroken\", or modified in a way that compromises security. It is recommended that you reset your device to its original state before proceeding."),
+        "save": MessageLookupByLibrary.simpleMessage("להציל"),
         "scanInstructions":
             MessageLookupByLibrary.simpleMessage("סרוק כתובת ברקוד \n של נאנו"),
         "scanNFC": MessageLookupByLibrary.simpleMessage("שלח דרך NFC"),
@@ -796,7 +801,7 @@ class MessageLookup extends MessageLookupByLibrary {
         "share": MessageLookupByLibrary.simpleMessage("לַחֲלוֹק"),
         "shareApp": MessageLookupByLibrary.simpleMessage("שתף %1"),
         "shareAppText": MessageLookupByLibrary.simpleMessage(
-            "בדוק את% 1! ארנק נייד NANO מעולה!"),
+            "בדוק את%1! ארנק נייד NANO מעולה!"),
         "shareLink": MessageLookupByLibrary.simpleMessage("שתף קישור"),
         "shareMessage": MessageLookupByLibrary.simpleMessage("שתף הודעה"),
         "shareNautilus": MessageLookupByLibrary.simpleMessage("Share Nautilus"),
@@ -825,7 +830,7 @@ class MessageLookup extends MessageLookupByLibrary {
         "simplex": MessageLookupByLibrary.simpleMessage("סימפלקס"),
         "social": MessageLookupByLibrary.simpleMessage("חֶברָתִי"),
         "someone": MessageLookupByLibrary.simpleMessage("מִישֶׁהוּ"),
-        "spendCurrency": MessageLookupByLibrary.simpleMessage("הוצא% 2"),
+        "spendCurrency": MessageLookupByLibrary.simpleMessage("הוצא%2"),
         "spendNano": MessageLookupByLibrary.simpleMessage("להוציא ננו"),
         "splitBill": MessageLookupByLibrary.simpleMessage("ביל מפוצל"),
         "splitBillHeader": MessageLookupByLibrary.simpleMessage("פיצול שטר"),
@@ -925,8 +930,8 @@ class MessageLookup extends MessageLookupByLibrary {
         "upcomingButton": MessageLookupByLibrary.simpleMessage("בקרוב"),
         "uptime": MessageLookupByLibrary.simpleMessage("זמן אונליין"),
         "urlEmpty": MessageLookupByLibrary.simpleMessage("נא להזין כתובת URL"),
-        "useAppRep": MessageLookupByLibrary.simpleMessage("השתמש ב-% 1 נציג"),
-        "useCurrency": MessageLookupByLibrary.simpleMessage("השתמש ב% 2"),
+        "useAppRep": MessageLookupByLibrary.simpleMessage("השתמש ב- 1 נציג"),
+        "useCurrency": MessageLookupByLibrary.simpleMessage("השתמש ב%2"),
         "useNano": MessageLookupByLibrary.simpleMessage("השתמש ב-NANO"),
         "useNautilusRep":
             MessageLookupByLibrary.simpleMessage("Use Nautilus Rep"),
@@ -962,6 +967,7 @@ class MessageLookup extends MessageLookupByLibrary {
         "watchOnlySendDisabled": MessageLookupByLibrary.simpleMessage(
             "השליחה מושבתת בכתובות לשעון בלבד"),
         "weekAgo": MessageLookupByLibrary.simpleMessage("לפני שבוע"),
+        "weekly": MessageLookupByLibrary.simpleMessage("שְׁבוּעִי"),
         "welcomeText": MessageLookupByLibrary.simpleMessage(
             "ברוך הבא לנאטריום. בכדי להמשיך, אתה רשאי לייצר ארנק חדש או לייבא אחד קיים."),
         "welcomeTextLogin": MessageLookupByLibrary.simpleMessage(
