@@ -132,7 +132,7 @@ class AppState extends State<App> {
             brightness: StateContainer.of(context).curTheme.brightness,
             error: StateContainer.of(context).curTheme.error,
             primary: StateContainer.of(context).curTheme.primary,
-            background: StateContainer.of(context).curTheme.background,
+            background: StateContainer.of(context).curTheme.backgroundDark,
             tertiary: StateContainer.of(context).curTheme.warning,
           ),
           // dialogBackgroundColor: StateContainer.of(context).curTheme.backgroundDarkest,
@@ -209,105 +209,7 @@ class AppState extends State<App> {
         locale: StateContainer.of(context).curLanguage.language == AvailableLanguage.DEFAULT
             ? null
             : StateContainer.of(context).curLanguage.getLocale(),
-        // supportedLocales: Z.delegate.supportedLocales,
-        supportedLocales: const <Locale>[
-          Locale('en', 'US'), // English
-          Locale('he', 'IL'), // Hebrew
-          Locale('de', 'DE'), // German
-          Locale('da'), // Danish
-          Locale('bg'), // Bulgarian
-          Locale('es'), // Spanish
-          Locale('hi'), // Hindi
-          Locale('hu'), // Hungarian
-          Locale('hi'), // Hindi
-          Locale('id'), // Indonesian
-          Locale('it'), // Italian
-          Locale('ja'), // Japanese
-          Locale('ko'), // Korean
-          Locale('ms'), // Malay
-          Locale('nl'), // Dutch
-          Locale('pl'), // Polish
-          Locale('pt'), // Portugese
-          Locale('ro'), // Romanian
-          Locale('ru'), // Russian
-          Locale('sl'), // Slovenian
-          Locale('sv'), // Swedish
-          Locale('tl'), // Tagalog
-          Locale('tr'), // Turkish
-          Locale('vi'), // Vietnamese
-          Locale('ca'), // Catalan
-          Locale('uk'), // Ukrainian
-          Locale('no'), // Norwegian
-          Locale.fromSubtags(languageCode: 'zh', scriptCode: 'Hans'), // Chinese Simplified
-          Locale.fromSubtags(languageCode: 'zh', scriptCode: 'Hant'), // Chinese Traditional
-          Locale('ar'), // Arabic
-          Locale('lv'), // Latvian
-          Locale('bn'), // Bengali
-          // Currency-default requires country included
-          Locale("es", "AR"),
-          Locale("en", "AU"),
-          Locale("pt", "BR"),
-          Locale("en", "CA"),
-          Locale("de", "CH"),
-          Locale("es", "CL"),
-          Locale("zh", "CN"),
-          Locale("cs", "CZ"),
-          Locale("da", "DK"),
-          Locale("fr", "FR"),
-          Locale("en", "GB"),
-          Locale("zh", "HK"),
-          Locale("hu", "HU"),
-          Locale("id", "ID"),
-          Locale("he", "IL"),
-          Locale("hi", "IN"),
-          Locale("ja", "JP"),
-          Locale("ko", "KR"),
-          Locale("es", "MX"),
-          Locale("ta", "MY"),
-          Locale("en", "NZ"),
-          Locale("tl", "PH"),
-          Locale("ur", "PK"),
-          Locale("pl", "PL"),
-          Locale("ru", "RU"),
-          Locale("sv", "SE"),
-          Locale("zh", "SG"),
-          Locale("th", "TH"),
-          Locale("tr", "TR"),
-          Locale("en", "TW"),
-          Locale("es", "VE"),
-          Locale("en", "ZA"),
-          Locale("en", "US"),
-          Locale("es", "AR"),
-          Locale("de", "AT"),
-          Locale("fr", "BE"),
-          Locale("de", "BE"),
-          Locale("nl", "BE"),
-          Locale("tr", "CY"),
-          Locale("et", "EE"),
-          Locale("fi", "FI"),
-          Locale("fr", "FR"),
-          Locale("el", "GR"),
-          Locale("es", "AR"),
-          Locale("en", "IE"),
-          Locale("it", "IT"),
-          Locale("es", "AR"),
-          Locale("lv", "LV"),
-          Locale("lt", "LT"),
-          Locale("fr", "LU"),
-          Locale("en", "MT"),
-          Locale("nl", "NL"),
-          Locale("pt", "PT"),
-          Locale("sk", "SK"),
-          Locale("sl", "SI"),
-          Locale("es", "ES"),
-          Locale("ar", "AE"), // UAE
-          Locale("ar", "SA"), // Saudi Arabia
-          Locale("ar", "KW"), // Kuwait
-          Locale("uk", "UA"), // Ukraine
-          Locale("no", "NO"), // Norway
-          Locale("bn", "BD"), // Bangladesh
-          Locale("bn", "IN"), // India/Bengali
-        ],
+        supportedLocales: Z.delegate.supportedLocales,
         initialRoute: '/',
         onGenerateRoute: (RouteSettings settings) {
           switch (settings.name) {
