@@ -211,15 +211,23 @@ class _RequestConfirmSheetState extends State<RequestConfirmSheet> {
             //A container for CONFIRM and CANCEL buttons
             Column(
               children: <Widget>[
-                // A row for CONFIRM Button
+
                 Row(
                   children: <Widget>[
+                    // CANCEL Button
+                    AppButton.buildAppButton(
+                        context,
+                        AppButtonType.PRIMARY_OUTLINE,
+                        CaseChange.toUpperCase(Z.of(context).cancel, context),
+                        Dimens.BUTTON_COMPACT_LEFT_DIMENS, onPressed: () {
+                      Navigator.of(context).pop();
+                    }),
                     // CONFIRM Button
                     AppButton.buildAppButton(
                         context,
                         AppButtonType.PRIMARY,
                         CaseChange.toUpperCase(Z.of(context).confirm, context),
-                        Dimens.BUTTON_TOP_DIMENS, onPressed: () async {
+                        Dimens.BUTTON_COMPACT_RIGHT_DIMENS, onPressed: () async {
                       if (clicking) return;
                       clicking = true;
                       // Authenticate
@@ -249,19 +257,6 @@ class _RequestConfirmSheetState extends State<RequestConfirmSheet> {
                       }
                       clicking = false;
                     })
-                  ],
-                ),
-                // A row for CANCEL Button
-                Row(
-                  children: <Widget>[
-                    // CANCEL Button
-                    AppButton.buildAppButton(
-                        context,
-                        AppButtonType.PRIMARY_OUTLINE,
-                        CaseChange.toUpperCase(Z.of(context).cancel, context),
-                        Dimens.BUTTON_BOTTOM_DIMENS, onPressed: () {
-                      Navigator.of(context).pop();
-                    }),
                   ],
                 ),
               ],
