@@ -27,7 +27,6 @@ import 'package:wallet_flutter/model/db/appdb.dart';
 import 'package:wallet_flutter/model/db/subscription.dart';
 import 'package:wallet_flutter/model/db/user.dart';
 import 'package:wallet_flutter/model/notification_setting.dart';
-import 'package:wallet_flutter/network/account_service.dart';
 import 'package:wallet_flutter/network/model/response/auth_item.dart';
 import 'package:wallet_flutter/network/model/response/pay_item.dart';
 import 'package:wallet_flutter/network/model/response/sub_item.dart';
@@ -37,9 +36,7 @@ import 'package:wallet_flutter/styles.dart';
 import 'package:wallet_flutter/ui/auth/auth_confirm_sheet.dart';
 import 'package:wallet_flutter/ui/handoff/handoff_confirm_sheet.dart';
 import 'package:wallet_flutter/ui/receive/receive_sheet.dart';
-import 'package:wallet_flutter/ui/scan/scan_screen.dart';
 import 'package:wallet_flutter/ui/send/send_confirm_sheet.dart';
-import 'package:wallet_flutter/ui/send/send_gift.dart';
 import 'package:wallet_flutter/ui/subs/sub_confirm_sheet.dart';
 import 'package:wallet_flutter/ui/util/confirm_sheet.dart';
 import 'package:wallet_flutter/ui/util/formatters.dart';
@@ -478,7 +475,9 @@ class SendSheetState extends State<SendSheet> {
           return;
         }
         // check if UD / ENS / opencap / onchain address:
-        if (_addressController.text.isNotEmpty && !_addressController.text.contains("★")) {
+        if (_addressController.text.isNotEmpty &&
+            !_addressController.text.contains("★") &&
+            !_addressController.text.startsWith(NonTranslatable.currencyPrefix)) {
           User? user = await sl.get<DBHelper>().getUserOrContactWithName(_addressController.text);
           if (user == null) {
             if (!mounted) return;
