@@ -439,6 +439,8 @@ String getXMRThemeAwareRawAccuracy(BuildContext context, String? raw) {
 String getThemeAwareAmountAsRaw(BuildContext context, String amount) {
   if (StateContainer.of(context).nyanoMode) {
     return NumberUtil.getNyanoAmountAsRaw(amount);
+  } else if (StateContainer.of(context).bananoMode) {
+    return NumberUtil.getBananoAmountAsRaw(amount);
   } else {
     return NumberUtil.getAmountAsRaw(amount);
   }
@@ -507,7 +509,9 @@ String getThemeCurrencyMode(BuildContext context) {
 // card time format:
 // String CARD_TIME_FORMAT = "MMM d, h:mm a";
 // ignore: non_constant_identifier_names
-String CARD_TIME_FORMAT = "MMM dd, HH:mm";
+// String CARD_TIME_FORMAT = "MMM dd, HH:mm";
+String OLD_CARD_TIME_FORMAT = "MMM dd, HH:mm";
+String CARD_TIME_FORMAT = "y/MM/dd, HH:mm";
 String CARD_HOUR_TIME_FORMAT = "HH:mm";
 String getTimeAgoString(BuildContext context, int epochTime) {
   String timeStr = DateFormat(CARD_TIME_FORMAT).format(DateTime.fromMillisecondsSinceEpoch(epochTime * 1000));
@@ -571,5 +575,10 @@ String getTimeAgoString(BuildContext context, int epochTime) {
 
 String getCardTime(int epochTime) {
   final String timeStr = DateFormat(CARD_TIME_FORMAT).format(DateTime.fromMillisecondsSinceEpoch(epochTime * 1000));
+  return timeStr;
+}
+
+String getOldCardTime(int epochTime) {
+  final String timeStr = DateFormat(OLD_CARD_TIME_FORMAT).format(DateTime.fromMillisecondsSinceEpoch(epochTime * 1000));
   return timeStr;
 }
