@@ -25,13 +25,15 @@ class MyInAppBrowser extends InAppBrowser {
     if (await canLaunchUrl(url)) {
       await launchUrl(url);
     }
-    return CustomSchemeResponse(contentType: "text/html", data: <int>[0] as Uint8List);
+    return CustomSchemeResponse(
+        contentType: "text/html", data: <int>[0] as Uint8List);
   }
 }
 
 class UIUtil {
   static Widget threeLineAddressText(BuildContext context, String address,
-      {ThreeLineAddressTextType type = ThreeLineAddressTextType.PRIMARY, String? contactName}) {
+      {ThreeLineAddressTextType type = ThreeLineAddressTextType.PRIMARY,
+      String? contactName}) {
     final String stringPartOne = address.substring(0, 12);
     final String stringPartTwo = address.substring(12, 22);
     final String stringPartThree = address.substring(22, 44);
@@ -85,7 +87,9 @@ class UIUtil {
                     text: stringPartFour,
                     style: AppStyles.textStyleAddressText60(context),
                   ),
-                  TextSpan(text: stringPartFive, style: AppStyles.textStyleAddressPrimary60(context)),
+                  TextSpan(
+                      text: stringPartFive,
+                      style: AppStyles.textStyleAddressPrimary60(context)),
                 ],
               ),
             )
@@ -94,7 +98,11 @@ class UIUtil {
       case ThreeLineAddressTextType.PRIMARY:
         Widget contactWidget;
         if (contactName != null) {
-          contactWidget = RichText(textAlign: TextAlign.center, text: TextSpan(text: contactName, style: AppStyles.textStyleAddressPrimary(context)));
+          contactWidget = RichText(
+              textAlign: TextAlign.center,
+              text: TextSpan(
+                  text: contactName,
+                  style: AppStyles.textStyleAddressPrimary(context)));
         } else {
           contactWidget = const SizedBox();
         }
@@ -150,7 +158,11 @@ class UIUtil {
       case ThreeLineAddressTextType.SUCCESS:
         Widget contactWidget;
         if (contactName != null) {
-          contactWidget = RichText(textAlign: TextAlign.center, text: TextSpan(text: contactName, style: AppStyles.textStyleAddressSuccess(context)));
+          contactWidget = RichText(
+              textAlign: TextAlign.center,
+              text: TextSpan(
+                  text: contactName,
+                  style: AppStyles.textStyleAddressSuccess(context)));
         } else {
           contactWidget = const SizedBox();
         }
@@ -257,7 +269,8 @@ class UIUtil {
     }
   }
 
-  static Widget oneLineAddressText(BuildContext context, String address, {OneLineAddressTextType type = OneLineAddressTextType.PRIMARY}) {
+  static Widget oneLineAddressText(BuildContext context, String address,
+      {OneLineAddressTextType type = OneLineAddressTextType.PRIMARY}) {
     final String stringPartOne = address.substring(0, 12);
     final String stringPartFive = address.substring(59);
     switch (type) {
@@ -341,7 +354,8 @@ class UIUtil {
     }
   }
 
-  static Widget threeLineSeedText(BuildContext context, String address, {TextStyle? textStyle}) {
+  static Widget threeLineSeedText(BuildContext context, String address,
+      {TextStyle? textStyle}) {
     textStyle = textStyle ?? AppStyles.textStyleSeed(context);
     final String stringPartOne = address.substring(0, 22);
     final String stringPartTwo = address.substring(22, 44);
@@ -364,7 +378,8 @@ class UIUtil {
     );
   }
 
-  static Widget sixLineSeedText(BuildContext context, String address, {TextStyle? textStyle}) {
+  static Widget sixLineSeedText(BuildContext context, String address,
+      {TextStyle? textStyle}) {
     textStyle = textStyle ?? AppStyles.textStyleSeed(context);
     final String stringPartOne = address.substring(0, 22);
     final String stringPartTwo = address.substring(22, 44);
@@ -402,7 +417,8 @@ class UIUtil {
     );
   }
 
-  static Future<void> showBlockExplorerWebview(BuildContext context, String? hash) async {
+  static Future<void> showBlockExplorerWebview(
+      BuildContext context, String? hash) async {
     cancelLockEvent();
     // final InAppBrowser browser = InAppBrowser();
     // final InAppBrowserClassOptions options = InAppBrowserClassOptions(
@@ -411,12 +427,14 @@ class UIUtil {
     //       toolbarTopBackgroundColor: StateContainer.of(context).curTheme.primary,
     //     ),
     //     inAppWebViewGroupOptions: InAppWebViewGroupOptions(crossPlatform: InAppWebViewOptions(javaScriptEnabled: true)));
-    final String url = NonTranslatable.getBlockExplorerUrl(hash, StateContainer.of(context).curBlockExplorer);
+    final String url = NonTranslatable.getBlockExplorerUrl(
+        hash, StateContainer.of(context).curBlockExplorer);
     // await browser.openUrlRequest(urlRequest: URLRequest(url: Uri.parse(url)), options: options);
     showChromeSafariWebview(context, url);
   }
 
-  static Future<void> showAccountWebview(BuildContext context, String? account) async {
+  static Future<void> showAccountWebview(
+      BuildContext context, String? account) async {
     cancelLockEvent();
     // final InAppBrowser browser = InAppBrowser();
     // final InAppBrowserClassOptions options = InAppBrowserClassOptions(
@@ -425,7 +443,8 @@ class UIUtil {
     //       toolbarTopBackgroundColor: StateContainer.of(context).curTheme.primary,
     //     ),
     //     inAppWebViewGroupOptions: InAppWebViewGroupOptions(crossPlatform: InAppWebViewOptions(javaScriptEnabled: true)));
-    final String url = NonTranslatable.getAccountExplorerUrl(account, StateContainer.of(context).curBlockExplorer);
+    final String url = NonTranslatable.getAccountExplorerUrl(
+        account, StateContainer.of(context).curBlockExplorer);
     // await browser.openUrlRequest(urlRequest: URLRequest(url: Uri.parse(url)), options: options);
     showChromeSafariWebview(context, url);
   }
@@ -443,20 +462,30 @@ class UIUtil {
         crossPlatform: InAppWebViewOptions(
           javaScriptEnabled: true,
           cacheEnabled: true,
-          resourceCustomSchemes: ["nano", "nanopay", "nanoauth", "nanosub", "nautilus"],
+          resourceCustomSchemes: [
+            "nano",
+            "nanopay",
+            "nanoauth",
+            "nanosub",
+            "nautilus"
+          ],
           useShouldOverrideUrlLoading: true,
         ),
       ),
     );
-    await browser.openUrlRequest(urlRequest: URLRequest(url: Uri.parse(url)), options: options);
+    await browser.openUrlRequest(
+        urlRequest: URLRequest(url: Uri.parse(url)), options: options);
   }
 
-  static Future<void> showChromeSafariWebview(BuildContext context, String url) async {
+  static Future<void> showChromeSafariWebview(
+      BuildContext context, String url) async {
     cancelLockEvent();
     final ChromeSafariBrowser browser = ChromeSafariBrowser();
 
-    final ChromeSafariBrowserClassOptions options = ChromeSafariBrowserClassOptions(
-      android: AndroidChromeCustomTabsOptions(shareState: CustomTabsShareState.SHARE_STATE_OFF),
+    final ChromeSafariBrowserClassOptions options =
+        ChromeSafariBrowserClassOptions(
+      android: AndroidChromeCustomTabsOptions(
+          shareState: CustomTabsShareState.SHARE_STATE_OFF),
       ios: IOSSafariOptions(barCollapsingEnabled: true),
     );
     await browser.open(url: Uri.parse(url), options: options);
@@ -508,29 +537,76 @@ class UIUtil {
     painter.paint(canvas, Size(qrSize, qrSize));
     final ui.Picture pic = recorder.endRecording();
     final ui.Image image = await pic.toImage(qrSize.toInt(), qrSize.toInt());
-    final ByteData? byteData = await image.toByteData(format: ui.ImageByteFormat.png);
+    final ByteData? byteData =
+        await image.toByteData(format: ui.ImageByteFormat.png);
     return Image.memory(byteData!.buffer.asUint8List());
   }
 
-  static void showSnackbar(String content, BuildContext context, {int durationMs = 2500}) {
+  static void showSnackbar(String content, BuildContext context,
+      {int durationMs = 2500, double topMarginPercent = 0.05}) {
     showToastWidget(
-      Align(
-        alignment: Alignment.topCenter,
-        child: Container(
-          margin: EdgeInsets.symmetric(vertical: MediaQuery.of(context).size.height * 0.05, horizontal: 14),
-          padding: const EdgeInsets.symmetric(vertical: 15, horizontal: 15),
-          width: MediaQuery.of(context).size.width - 30,
-          decoration: BoxDecoration(
-            color: StateContainer.of(context).curTheme.primary,
-            borderRadius: BorderRadius.circular(10),
-            boxShadow: [
-              BoxShadow(color: StateContainer.of(context).curTheme.barrier!, offset: const Offset(0, 15), blurRadius: 30, spreadRadius: -5),
-            ],
+      SafeArea(
+        child: Align(
+          alignment: Alignment.topCenter,
+          child: Container(
+            // margin: EdgeInsets.symmetric(vertical: MediaQuery.of(context).size.height * 0.07, horizontal: 14),
+            margin: EdgeInsets.only(
+                top: MediaQuery.of(context).size.height * topMarginPercent,
+                left: 14,
+                right: 14),
+            padding: const EdgeInsets.symmetric(vertical: 15, horizontal: 15),
+            width: MediaQuery.of(context).size.width - 30,
+            decoration: BoxDecoration(
+              color: StateContainer.of(context).curTheme.warning,
+              borderRadius: BorderRadius.circular(10),
+              boxShadow: [
+                BoxShadow(
+                    color: StateContainer.of(context).curTheme.barrier!,
+                    offset: const Offset(0, 15),
+                    blurRadius: 30,
+                    spreadRadius: -5),
+              ],
+            ),
+            child: Text(
+              content,
+              style: AppStyles.textStyleSnackbar(context),
+              textAlign: TextAlign.start,
+            ),
           ),
-          child: Text(
-            content,
-            style: AppStyles.textStyleSnackbar(context),
-            textAlign: TextAlign.start,
+        ),
+      ),
+      dismissOtherToast: true,
+      duration: Duration(milliseconds: durationMs),
+    );
+  }
+
+  static void showSnackbarNoContext(String content,
+      {int durationMs = 2500, double topMarginPercent = 0.05}) {
+    showToastWidget(
+      SafeArea(
+        child: Align(
+          alignment: Alignment.topCenter,
+          child: Container(
+            // margin: EdgeInsets.symmetric(vertical: MediaQuery.of(context).size.height * 0.07, horizontal: 14),
+            margin: EdgeInsets.only(top: 40, left: 14, right: 14),
+            padding: const EdgeInsets.symmetric(vertical: 15, horizontal: 15),
+            // width: MediaQuery.of(context).size.width - 30,
+            decoration: BoxDecoration(
+              // color: StateContainer.of(context).curTheme.warning,
+              borderRadius: BorderRadius.circular(10),
+              // boxShadow: [
+              //   BoxShadow(
+              //       // color: StateContainer.of(context).curTheme.barrier!,
+              //       offset: const Offset(0, 15),
+              //       blurRadius: 30,
+              //       spreadRadius: -5),
+              // ],
+            ),
+            child: Text(
+              content,
+              // style: AppStyles.textStyleSnackbar(context),
+              textAlign: TextAlign.start,
+            ),
           ),
         ),
       ),
@@ -564,11 +640,12 @@ class UIUtil {
   }
 }
 
-
 /// This is used so that the elevation of the container is kept and the
 /// drop shadow is not clipped.
 class SizeTransitionNoClip extends AnimatedWidget {
-  const SizeTransitionNoClip({required Animation<double> sizeFactor, this.child}) : super(listenable: sizeFactor);
+  const SizeTransitionNoClip(
+      {required Animation<double> sizeFactor, this.child})
+      : super(listenable: sizeFactor);
 
   final Widget? child;
 
